+++ header_test.go
@@ -130,69 +130,3 @@ func TestWriteThenReadHeader(t *testing.T) {
 		})
 	}
 }
-
-type testCompressionSpec struct {
-	SpecName                  string
-	Compression               Compression
-	CompressionCodec          CompressionCodec
-	CompressionCodecClassName string
-}
-
-var testcompressionspecs = []testCompressionSpec{
-	{
-		SpecName:                  "NoCompression",
-		Compression:               NoCompression,
-		CompressionCodec:          SnappyCompression,
-		CompressionCodecClassName: SnappyClassName,
-	},
-	{
-		SpecName:                  "BlockCompression with SnappyCompression",
-		Compression:               BlockCompression,
-		CompressionCodec:          SnappyCompression,
-		CompressionCodecClassName: SnappyClassName,
-	},
-	{
-		SpecName:                  "RecordCompression with SnappyCompression",
-		Compression:               RecordCompression,
-		CompressionCodec:          SnappyCompression,
-		CompressionCodecClassName: SnappyClassName,
-	},
-	{
-		SpecName:                  "BlockCompression with GzipCompression",
-		Compression:               BlockCompression,
-		CompressionCodec:          GzipCompression,
-		CompressionCodecClassName: GzipClassName,
-	},
-	{
-		SpecName:                  "RecordCompression with GzipCompression",
-		Compression:               RecordCompression,
-		CompressionCodec:          GzipCompression,
-		CompressionCodecClassName: GzipClassName,
-	},
-}
-
-func TestWriteHeaderCompression(t *testing.T) {
-	for _, spec := range testcompressionspecs {
-		t.Run(spec.SpecName, func(t *testing.T) {
-			buf := new(bytes.Buffer)
-			writer := NewWriter(buf)
-
-			writer.Header.Compression = spec.Compression
-			writer.Header.CompressionCodec = spec.CompressionCodec
-
-			_, err := writer.WriteHeader()
-			assert.NoError(t, err, "it should write successfully")
-
-			r := NewReader(buf)
-			err = r.ReadHeader()
-			assert.NoError(t, err, "it should read successfully")
-
-			assert.Equal(t, spec.Compression, r.Header.Compression, "it should have the correct compression type")
-			if spec.Compression != NoCompression {
-				assert.Equal(t, spec.CompressionCodecClassName, r.Header.CompressionCodecClassName, "it should have the correct compression codec class name")
-				assert.Equal(t, spec.CompressionCodec, r.Header.CompressionCodec, "it should have the correct compression codec")
-			}
-		})
-	}
-
-}
